Address PR review feedback
- Replace @since 0.6.0 with @since x.x.x placeholders in bootstrap.php
- Add /routes and /tools to .distignore
- Keep Settings_Page class with static init() method instead of inlining admin_menu logic in bootstrap.php
- Extend webpack clean.keep regex to preserve wp-build output directories during watch mode
